move command check helper functions out of main test index
